Move the "create meta" method to different files
