+++ common/src/codegen/index.ts
@@ -1,5 +1,7 @@
-import { createOutputProjectMetadata, createSourceProjectMetadata, generateCode } from "./codegen";
+import { generateCode } from "./codegen";
 import { CodeGenMetaFactory } from "./codegenMetaFactory";
+import { createOutputProjectMetadata } from "./codegenOutputProjectMeta";
+import { createSourceProjectMetadata } from "./codegenSourceProjectMeta";
 import { IOutputProjectMetadataProcessor } from "./playwright-charp-common/outputProjectMetadataProcessor";
 import { ICodeGen } from "./types";
