+++ common/src/codegen/playwright-csharp-nunit/playwrightCsharpNunitCodeGen.ts
@@ -11,7 +11,6 @@ import {
 } from "../../file-defs";
 import { IPage } from "../../file-defs/pageFile";
 import { StandardOutputFile } from "../../file-defs/standardOutputFile";
-import { createOutputProjectMetadata } from "../codegen";
 import { ICodeGen } from "../types";
 import { addIndent, hasPlaceholder, upperCaseFirstChar } from "../utils/stringUtils";
 import { NunitProjectMeta } from "./nunitProjectMeta";
@@ -21,6 +20,7 @@ import { PlaywrightCsharpCodeGen } from "../playwright-charp-common/playwrightCs
 import { IOutputProjectMetadataProcessor } from "../playwright-charp-common/outputProjectMetadataProcessor";
 import generateDatasetInfos from "../playwright-charp-common/generateDatasetInfos";
 import { IPlaywrightCsharpTemplatesProvider } from "../playwright-charp-common/playwrightCsharpTemplatesProvider";
+import { createOutputProjectMetadata } from "../codegenOutputProjectMeta";
 
 export class PlaywrightCsharpNunitCodeGen extends PlaywrightCsharpCodeGen implements ICodeGen {
   constructor(projMeta: ISourceProjectMetadata) {
